[PIDM-1763] change amount type from Integer to Long (payment-methods-service) (#301)
* fix: change amount type from integer to long

* chore: disable depcheck

* chore: add maven-scm-plugin

* chore: add maven-scm-plugin

* chore: use <pagopa-ecommerce-commons.version>3.3.1-rc</pagopa-ecommerce-commons.version>

* chore: add validate  in goals

* chore: add new step in codereview pipeline (Bootstrap pagopa-ecommerce-commons)

* chore: add new step in codereview pipeline (Bootstrap pagopa-ecommerce-commons)

* chore: add new step in codereview pipeline (Bootstrap pagopa-ecommerce-commons)

* chore: add new step in codereview pipeline (Bootstrap pagopa-ecommerce-commons)

* chore: testing

* chore: testing

* chore: testing

* chore: testing

* chore: testing

* chore: testing

* chore: testing

* chore: testing

* chore: testing

* chore: testing

* test: set version >pagopa-ecommerce-commons -> 3.3.2-PIDM-1763-amount-int-to-long-SNAPSHOT-b5d38e4

* test: set version >pagopa-ecommerce-commons -> 3.3.2-PIDM-1763-amount-int-to-long-SNAPSHOT-b5d38e4

* chore: remove maven-scm-plugin

* chore: restore code-review pipeline

* chore: set ${pagopa-ecommerce-commons.version}

* chore: set pagopa-ecommerce-commons.version  3.3.2-PIDM-1763-amount-int-to-long-SNAPSHOT-19278cb

* fix: Fix commons version 3.5.0

* fix: enable depcheck plugin

* fix: set 		<pagopa-ecommerce-commons.version>3.6.0</pagopa-ecommerce-commons.version>

* fix: update dep-sha256.json

* test: test version commons 3.6.0-PIDM-1763-grandTotal-Long-SNAPSHOT-a7f6848

* test: test version commons 3.6.0-PIDM-1763-grandTotal-Long-SNAPSHOT-a7f6848

* fix: update commons version 3.6.1

---------

+++ src/main/java/it/pagopa/ecommerce/payment/methods/controller/v1/PaymentMethodsController.java
@@ -146,7 +146,7 @@ public class PaymentMethodsController implements PaymentMethodsApi {
                                                                                 ServerWebExchange exchange
     ) {
         return paymentMethodService
-                .retrievePaymentMethods(amount != null ? amount.intValue() : null, xClientId, deviceVersion)
+                .retrievePaymentMethods(amount != null ? amount.longValue() : null, xClientId, deviceVersion)
                 .map(PaymentMethodsController::paymentMethodToDto)
                 .collectList()
                 .map(
